chore(provision): extract comparison helpers, fix CLI parser, add regression coverage
- Extract canonical normalization, semantic comparison, and divergence helpers from invitation-import-engine.ts into promotion-comparison.ts, reducing import-engine from 988 to 774 lines (below 800 max-lines lint)
- Extract buildResourceActions into the same module via import type
- Fix finalize-preview-approval-cli.ts argument parser: guard indexOf -1 so absent --reject does not read args[0] as reason
- Export parseFinalizerArgs and add 6-unit regression test suite
- Update test imports for moved functions
- Restore approval artifact to honest pending_hosted_validation state
